contract_recurring_plans
==============================================
Analytic plans for contract recurring invoices
==============================================

This module allows to include an analytic plan on each recurring invoice line,
so if you set it, this will be taken to be transferred to the invoice instead
of the analytic account of the contract.

Usage
=====

On the contract view, select or create an analytic distribution for the
invoice recurring lines that you want.