[ADD] contract_recurring_invoicing_markers
=======================================
Markers for contract recurring invoices
=======================================

This module allows to include some markers on the lines of your recurring
invoices definition inside the contract so that the generated invoice lines
will contain a dynamic text depending on these markers.

These markers are the supported ones:

* #START#: Start date of the invoiced period.
* #END# End date of the invoiced period.

Usage
=====

On a contract (*Sales > Sales > Contracts*), mark "Generate recurring invoices
automatically" for enabling the creation of recurring invoices.

In the "Invoice Lines" section, you can add lines with products. In the
*Description* field, you can now add any of the markers mentioned before
between the rest of the text.

When you invoice this contract (automatically or manually), your invoice
will contain the corresponding text that replaces the marker.

Known issues / Roadmap
======================

* Add more markers, like #START_MONTH# or #END_MONTH#.