[REM] contract_discount/contrac_journal/contract_show_recurring_invoice
These modules have been integrated in the new contract module, which is itself
an extraction of the recurring invoicing feature from Odoo's v8
account_analytic_analysis module.

As in v9 a similar option has been included in enterprise, but not in community
version, this migration to v9 will serve as base for the suscription functionality.