[FIX] contract: Process invoices limit in a lower method
To avoid blocking the queue when there are more than the specified limit
of contracts, we process the limit while creating invoices instead
of while searching for contracts, and break the process when the
max of invoices has been created.

See https://github.com/OCA/contract/pull/260#pullrequestreview-192187022
for more details.

In case you need to use this new feature in the cron, it is
also modified as `noupdate=1`.