[FIX] Un-optimize invoice creation to fix wrong auto-payment amount bug
Adding a test for the payment amount in contract_payment_auto makes it
visible that the usage of the `norecompute` context manager in contrat
`recurring_invoice` was breaking the contract_payment_auto module's
behaviour.

To fix this without changing the de facto API of this module
(_create_invoice, which is badly prefixed by an underscore), the
`norecompute context manager` was removed and the
`_onchange_invoice_line_ids` method of account.invoice model is now
explicitely called.

Moreover, the invoice creation context is now set around the
`_create_invoice` method call so that it can be used in the modules
that implement a custom payment method and want to benefit from that
data.