[FIX] contract: product is also required in contract line tree view
The Product is a mandatory field in the contract line form view

https://github.com/OCA/contract/blob/14.0/contract/views/abstract_contract_line.xml#L23

but it's not set as mandatory in the embedded tree view of the
contract line. This commit fixes this inconsistency.

This causes problems in particular with module contract_sale_generation,
since, unlike invoices, it's mandatory
to have a product on a Sales Order line.