[IMP] contract_sale_invoicing: improve search for Sale Order Lines with analytic account
A variable distribution of multiple analytic accounts per Sale Order Line is possible.
This change adapts the module to consider only Sale Order Lines that match 100%
with the analytic account of the contract.

Additionally, when a Sale Order Line does not have an analytic_distribution,
but the order has an analytic_account_id that matches the contract,
that line is also taken into account.