[IMP] contract_sale_invoicing: Add option to include sale lines in the same contract invoice
Before this commit, when a sale order was invoiced from a contract, it created a new invoice, resulting in two invoices, one for the contract and another for the sale order.
After this commit, when this new field is checked in the contract, the sale order lines are invoiced in the same invoice as the contract.