[FIX] contract: implement missing _compute_analytic_distribution on contract line
Adding analytic.mixin on a model adds the analytic_distribution compute field
In the mixin the compute is not implemented, letting the method being implemented
in each concrete model. This wasn't done yet for contract line.

Before this commit: adding a product to a contract line didn't fill the analytic distribution.

After this commit: adding a product to a contract line triggers the compute method
(similarly to the one on sale order lines), evaluating the account analytic distribution
models and setting the correct analytic distribution